server/requirements.txt
# Core web framework (Python 3.12 compatible versions)
fastapi>=0.100.0
pydantic[email]>=2.0.0
uvicorn>=0.23.0
starlette>=0.27.0

# Database
pymongo>=4.0.0

# Utilities
typing_extensions>=4.0.0
requests>=2.28.0
python-multipart
aiofiles
python-decouple
numpy>=1.24.0
pandas>=2.0.0

# Authentication
PyJWT>=2.2.0
pyopenssl>=23.0.0

# OCR and image processing
PyMuPDF>=1.23.0  # PDF to PNG conversion (standalone, no Poppler needed)
kraken>=6.0.0
opencv-python-headless
python-dotenv
httpx>=0.24.0
pygit2
pgzip

# YOLO for layout detection
ultralytics>=8.0.0

# eBL-ATF parsing (for local validation)
lark-parser>=0.11.0

# TEI Lex-0 XML validation
lxml>=5.0.0

# VLM OCR (optional - for DeepSeek-OCR)
# Need 4.46+ for tokenizers compatibility with Nemotron model
transformers>=4.46.0,<5.0.0

# Cloud VLM APIs
google-genai>=1.0.0
anthropic>=0.40.0
openai>=1.50.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
open_clip_torch  # Required for Nemotron-Parse vision encoder

# CuRe sign classifier (ResNet18)
torchvision>=0.15.0
scikit-learn>=1.3.0
